% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dareML.R
\docType{package}
\name{dareML}
\alias{dareML}
\alias{dareML-package}
\title{dareML: Library for easily obtain clean and real datasets for machine learning projects or practive, suitable for differente types of algorithms.}
\description{
Library for easily obtain clean and real datasets for machine learning projects or practive, suitable for differente types of algorithms.
}
\author{
Miguel Arquez \email{miguel.arquez12@gmail.com}
}
