% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance-datasets.R
\name{downloadStockData}
\alias{downloadStockData}
\title{downloadStockData from Quandl API}
\usage{
downloadStockData(stock, database = "WIKI", format = c("csv", "rds", "xlsx"))
}
\arguments{
\item{format}{}
}
\value{

}
\description{
downloadStockData from Quandl API
}
