% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-datasets.R
\name{getIris}
\alias{getIris}
\title{Call iris dataset}
\usage{
getIris()
}
\value{
\code{iris} dataset with the following columns:
\itemize{
\item Sepal.Length
\item Sepal.Width
\item Petal.Length
\item Petal.Width
\item Species
}
}
\description{
Ideal dataset for clustering, multi-class
problems (tree based models) and EDA.
}
\examples{
my_iris_df <- getIris()
}
\author{
Miguel Arquez
}
