% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci-datasets.R
\name{loadAirQuality}
\alias{loadAirQuality}
\title{airQuality time series dataset}
\usage{
loadAirQuality(mode = c("load", "download"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
airQuality time series dataset
}
