% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci-datasets.R
\name{loadBikeSharing}
\alias{loadBikeSharing}
\title{load Bike Sharing data}
\usage{
loadBikeSharing(mode = c("load", "download"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
load Bike Sharing data
}
