% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci-datasets.R
\name{loadDowJonesIndex}
\alias{loadDowJonesIndex}
\title{dowJones index Time Series Dataset}
\usage{
loadDowJonesIndex(mode = c("load", "download"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
dowJones index Time Series Dataset
}
