% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci-datasets.R
\name{loadGenderByName}
\alias{loadGenderByName}
\title{Gender by name - classification dataset}
\usage{
loadGenderByName(mode = c("load", "download"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Gender by name - classification dataset
}
