% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-datasets.R
\name{loadHousingDataSet}
\alias{loadHousingDataSet}
\title{Housing dataset loaded in current environment}
\usage{
loadHousingDataSet(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Housing dataset loaded in current environment
}
