% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-datasets.R
\name{loadIris}
\alias{loadIris}
\title{Load Iris dataset in the current environment}
\usage{
loadIris()
}
\value{

}
\description{
Load the iris dataset as a \code{data.frame} object called
\code{irisDF}
}
\examples{
loadIris()
}
