% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dareML.R
\docType{package}
\name{dareML}
\alias{dareML}
\alias{dareML-package}
\title{dareML: DAta REpository for Machine Learning}
\description{
@description  Library for easily obtain clean and real datasets
for machine learning projects or practice, suitable
for differente types of algorithms.
}
\author{
Miguel Arquez \email{miguel.arquez12@gmail.com}
}
